Gave the tabbedredux app the expected file structure as opposed of having everything in a single file.
This was prep work to get the tabs to work as expected.
