+++ samples/13.tabbedredux/src/containers/index.tsx
@@ -1,0 +1,5 @@
+import { RootNavigatorContainer, RootNavigator } from "./rootNavigator";
+import { ScreenOneContainer } from "./screenOne";
+import { ScreenTwo } from "./screenTwo";
+
+export { RootNavigatorContainer, RootNavigator, ScreenOneContainer, ScreenTwo };
