+++ samples/13.tabbedredux/src/routes.tsx
@@ -1,0 +1,9 @@
+/**
+ * With the screens defined, we now know there are two possible routes we can do.
+ */
+enum RouteNames {
+    One = "/one",
+    Two = "/two"
+}
+
+export { RouteNames };
